(PC-32696)[PRO] refactor: mv other constants in appropriate IndividualOffer folder
