--- pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferInformationsLegacy/InformationsScreen/IndividualOfferForm/Accessibility/constants.ts
@@ -1,11 +1,0 @@
-import { AccessibilityEnum } from 'commons/core/shared/types'
-
-export const ACCESSIBILITY_DEFAULT_VALUES = {
-  accessibility: {
-    [AccessibilityEnum.VISUAL]: false,
-    [AccessibilityEnum.MENTAL]: false,
-    [AccessibilityEnum.AUDIO]: false,
-    [AccessibilityEnum.MOTOR]: false,
-    [AccessibilityEnum.NONE]: false,
-  },
-}
