+++ pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferInformationsLegacy/InformationsScreen/IndividualOfferForm/Categories/ShowTypes/ShowTypes.tsx
@@ -3,8 +3,8 @@ import { useFormikContext } from 'formik'
 import { showOptionsTree } from 'commons/core/Offers/categoriesSubTypes'
 import { FormLayout } from 'components/FormLayout/FormLayout'
 import { IndividualOfferFormValues } from 'pages/IndividualOffer/commons/types'
+import { FORM_DEFAULT_VALUES } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/commons/constants'
 import { buildShowSubTypeOptions } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/utils/utils'
-import { FORM_DEFAULT_VALUES } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferInformationsLegacy/InformationsScreen/IndividualOfferForm/commons/constants'
 import { Select } from 'ui-kit/form/Select/Select'
 
 interface ShowTypesProps {
