--- pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferInformationsLegacy/InformationsScreen/IndividualOfferForm/Categories/constants.ts
@@ -1,9 +1,0 @@
-import { SUBCATEGORIES_FIELDS_DEFAULT_VALUES } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/commons/constants'
-
-export const CATEGORIES_DEFAULT_VALUES = {
-  subCategoryFields: [],
-  categoryId: '',
-  subcategoryId: '',
-  isEvent: false,
-  ...SUBCATEGORIES_FIELDS_DEFAULT_VALUES,
-}
