--- pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferInformationsLegacy/InformationsScreen/IndividualOfferForm/Notifications/constants.ts
@@ -1,4 +1,0 @@
-export const NOTIFICATIONS_DEFAULT_VALUES = {
-  receiveNotificationEmails: false,
-  bookingEmail: '',
-}
