+++ pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/commons/constants.ts
@@ -1,3 +1,5 @@
+import { AccessibilityEnum } from 'commons/core/shared/types'
+
 export const SUBCATEGORIES_FIELDS_DEFAULT_VALUES = {
   gtl_id: '',
   musicType: '',
@@ -13,4 +15,79 @@ export const SUBCATEGORIES_FIELDS_DEFAULT_VALUES = {
   durationMinutes: '',
   withdrawalType: undefined,
   withdrawalDelay: undefined,
+}
+
+const OFFER_LOCATION_DEFAULT_VALUES = {
+  locationLabel: '',
+  manuallySetAddress: false,
+  street: '',
+  'search-addressAutocomplete': '',
+  addressAutocomplete: '',
+  postalCode: '',
+  city: '',
+  coords: '',
+  latitude: '',
+  longitude: '',
+  banId: '',
+}
+
+const INFORMATIONS_DEFAULT_VALUES = {
+  name: '',
+  description: '',
+  author: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['author'],
+  ean: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['ean'],
+  performer: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['performer'],
+  speaker: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['speaker'],
+  stageDirector: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['stageDirector'],
+  visa: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['visa'],
+  durationMinutes: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['durationMinutes'],
+}
+
+export const CATEGORIES_DEFAULT_VALUES = {
+  subCategoryFields: [],
+  categoryId: '',
+  subcategoryId: '',
+  isEvent: false,
+  ...SUBCATEGORIES_FIELDS_DEFAULT_VALUES,
+}
+
+const USEFUL_INFORMATIONS_DEFAULT_VALUES = {
+  offererId: '',
+  venueId: '',
+  withdrawalDetails: '',
+  withdrawalType: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['withdrawalType'],
+  withdrawalDelay: SUBCATEGORIES_FIELDS_DEFAULT_VALUES['withdrawalDelay'],
+  isNational: false,
+  url: '',
+  bookingContact: '',
+}
+
+const ACCESSIBILITY_DEFAULT_VALUES = {
+  accessibility: {
+    [AccessibilityEnum.VISUAL]: false,
+    [AccessibilityEnum.MENTAL]: false,
+    [AccessibilityEnum.AUDIO]: false,
+    [AccessibilityEnum.MOTOR]: false,
+    [AccessibilityEnum.NONE]: false,
+  },
+}
+
+export const NOTIFICATIONS_DEFAULT_VALUES = {
+  receiveNotificationEmails: false,
+  bookingEmail: '',
+}
+
+export const FORM_DEFAULT_VALUES_NO_OFFER_LOCATION = {
+  ...INFORMATIONS_DEFAULT_VALUES,
+  ...CATEGORIES_DEFAULT_VALUES,
+  ...USEFUL_INFORMATIONS_DEFAULT_VALUES,
+  ...ACCESSIBILITY_DEFAULT_VALUES,
+  ...NOTIFICATIONS_DEFAULT_VALUES,
+  isDuo: false,
+}
+
+export const FORM_DEFAULT_VALUES = {
+  ...FORM_DEFAULT_VALUES_NO_OFFER_LOCATION,
+  ...OFFER_LOCATION_DEFAULT_VALUES,
+  isDuo: false,
 }
