--- pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/components/OfferLocation/constants.ts
@@ -1,13 +1,0 @@
-export const OFFER_LOCATION_DEFAULT_VALUES = {
-  locationLabel: '',
-  manuallySetAddress: false,
-  street: '',
-  'search-addressAutocomplete': '',
-  addressAutocomplete: '',
-  postalCode: '',
-  city: '',
-  coords: '',
-  latitude: '',
-  longitude: '',
-  banId: '',
-}
